+++ tests/TestCase.php
@@ -82,6 +82,7 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase {
                             'scriptUrl' => '/index.php',
                         ],
                         'user' => [
+                            'class' => '\vistart\Models\components\SSOIdentity',
                             'identityClass' => 'app\models\user\User',
                             'enableAutoLogin' => true,
                         ]
@@ -104,12 +105,13 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase {
             $this->markTestSkipped('No mysql server connection configured.');
         }
         $connection = new Connection($params);
+        $md = self::getParam('multipleDomainsManager');
         $redis = self::getParam('redis');
         $cacheParams = self::getParam('cache');/*
         if ($cacheParams === null) {
             $this->markTestSkipped('No cache component configured.');;
         }*/
-        $this->mockWebApplication(['components' => ['redis' => $redis, 'db' => $connection, 'cache' => $cacheParams]]);
+        $this->mockWebApplication(['components' => ['redis' => $redis, 'multipleDomainsManager' => $md, 'db' => $connection, 'cache' => $cacheParams]]);
 
         parent::setUp();
     }